Split global inbox watcher from rollup watcher
