+++ packages/arb-validator-core/arbbridge/arbRollupWatcher.go
@@ -31,3 +31,7 @@ type ArbRollupWatcher interface {
 	GetCreationInfo(ctx context.Context) (common.Hash, *common.BlockId, common.Hash, error)
 	GetVersion(ctx context.Context) (string, error)
 }
+
+type GlobalInboxWatcher interface {
+	ContractWatcher
+}
